requirements.txt
Flask==3.1.1
Pillow==11.3.0
python-pptx==1.0.2
Werkzeug==3.1.3
XlsxWriter==3.2.5
lxml==6.0.0
reportlab==4.2.2
aiohttp==3.12.15
python-dotenv==1.0.0
gunicorn==21.2.0
# Additional dependencies for serverless compatibility
requests==2.31.0
urllib3==2.0.7